+++ client/tests/unit/pages/calendar/utils.spec.js
@@ -11,6 +11,7 @@ describe('formatEvent', () => {
     start_date: '2019-10-01',
     end_date: '2019-10-02',
     is_confirmed: false,
+    has_missing_materials: false,
   };
 
   it('returns the event with moment dates and time flags', () => {
@@ -26,6 +27,8 @@ describe('formatEvent', () => {
       is_confirmed: false,
       isConfirmed: false,
       isCurrent: false,
+      has_missing_materials: false,
+      hasMissingMaterials: false,
       isPast: true,
     });
   });
@@ -38,11 +41,12 @@ describe('formatTimelineEvent', () => {
     start_date: '2019-10-01',
     end_date: '2019-10-02',
     is_confirmed: false,
+    has_missing_materials: true,
     location: 'Testville',
   };
 
   it('returns the event well formated to be used in timeline', () => {
-    const result = utils.formatTimelineEvent(event);
+    const result = utils.formatTimelineEvent(event, (s) => s);
     expect(result).toBeDefined();
     const resultWithStringDates = JSON.parse(JSON.stringify(result));
     expect(resultWithStringDates).toEqual({
@@ -51,7 +55,8 @@ describe('formatTimelineEvent', () => {
       start: '2019-10-01T00:00:00.000Z',
       end: '2019-10-02T00:00:00.000Z',
       editable: false,
-      className: 'Calendar__event Calendar__event--past Calendar__event--locked',
+      className: 'Calendar__event Calendar__event--past Calendar__event--locked Calendar__event--with-warning',
+      title: "Test (Testville)\n  →page-calendar.this-event-is-past\n  →page-calendar.this-event-has-missing-materials"
     });
   });
 });